move six calls / metaclass rewrites to plugins
